Minor fix
Fixed how data points are reset when the user is entering data for PDF(s). Previously all data points were reset to initial values, so anything the user had entered was lost. (And previous to that, if the user had created a PDF once or more and was creating another one, they might be asked for data points unneeded for the PDF(s) they were actually creating this time around.) Data points should be reset appropriately now.